fix: keep perps search aliases out of main startup
